# model-pricing-backend/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "大模型中转平台：模型数据中心 + 进货/出货两条定价线。统一响应包 {code,message,data,requestId}；金额一律字符串传输；时间 ISO 8601 带时区。\n错误码：400 参数校验、401 未认证、403 无权限、409 幂等/状态冲突、423 主体冻结、429 限流。\n写操作（产生业务新版本/影响资金合同）必须携带 Idempotency-Key 头。\n分页：?page=1\u0026size=20 或 ?cursor=\u0026size=，统一返回 {list,total,page,size}。",
        "title": "模型管理与定价中心 API",
        "contact": {
            "name": "后端"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/api/open/aliases": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "since 为客户端已见版本号；since \u003e= version 返回空 items，否则返回全量别名（联 model_sku 拿 sku_code）。不含 cost/margin/floor_price。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "开放接口"
                ],
                "summary": "模型别名列表（开放接口）",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "客户端已见版本号",
                        "name": "since",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data=AliasesResult{version,items[]}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 token 已过期或不存在",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/open/auth/token": {
            "post": {
                "description": "client_id + client_secret（sys_config 配置）换 1h 短期 token。与三大门户 login_session 完全隔离。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "开放接口"
                ],
                "summary": "开放接口 token 签发",
                "parameters": [
                    {
                        "description": "凭证",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "client_id": {
                                    "type": "string"
                                },
                                "client_secret": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data=TokenResult{token,expires_at}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "code=10001 参数非法",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 client_id 或 client_secret 错误",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/open/cost-snapshot": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回聚合 unit_cost（代表组件完全成本），无组件明细。asOf 为空=当前版本，非空=该时刻生效版本（valid_from\u003c=asOf AND (valid_to IS NULL OR valid_to\u003easOf)）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "开放接口"
                ],
                "summary": "成本快照（开放接口）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sku_id 或 sku_code",
                        "name": "sku",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 时刻",
                        "name": "asOf",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data=CostSnapshotResult{sku_id,unit_cost,currency,as_of,baseline_version}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "code=10001 sku 必填 / asOf 非法",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 token 已过期或不存在",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "code=10004 SKU 不存在 / 暂无成本基线",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/open/events": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "30s 超时，每秒 tick 一次。since 过滤 id \u003e since；limit 默认 100 上限 100。超时返回空列表（非错误）。payload 为 jsonb 全量透传。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "开放接口"
                ],
                "summary": "长轮询事件（开放接口）",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "客户端已见最大事件 id",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "返回上限",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data=EventsResult{last_id,events[]}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 token 已过期或不存在",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/open/price-book": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按 level_code 返回当前 EFFECTIVE 价目表（联 model_sku + 代表组件 unit_price）。不含 floor_price/unit_cost/margin/baseline。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "开放接口"
                ],
                "summary": "当前生效价目表（开放接口）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "价目表等级（如 GLOBAL）",
                        "name": "level",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data=PriceBookResult{level_code,version_no,currency,items[]}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "code=10001 level 必填",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 token 已过期或不存在",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "code=10004 该等级暂无生效价目表",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/open/routing/{sku}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "primary=当前成本基线主供应商（weight=100），backups=按四因子评分排序的其余供应商（weight=0）。无成本基线或无 EFFECTIVE 报价 → 404。不返回评分明细。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "开放接口"
                ],
                "summary": "主备路由（开放接口）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sku_id 或 sku_code",
                        "name": "sku",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data=RoutingResult{sku_id,primary,backups[],currency}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 token 已过期或不存在",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "code=10004 SKU 不存在 / 暂无成本基线",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/open/sellable-models": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回 lifecycle_status IN ('PUBLISHED','PURCHASABLE') 的 SKU 列表。level_tags 从 model_sku.tags 透传（NULL→[]）。不含 cost/margin/floor_price。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "开放接口"
                ],
                "summary": "上架可售模型（开放接口）",
                "responses": {
                    "200": {
                        "description": "code=0；data=SellableModelsResult{items[]}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 token 已过期或不存在",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务错误码，0=成功",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "业务数据",
                    "type": "object"
                },
                "message": {
                    "description": "错误消息，成功时 \"ok\"",
                    "type": "string",
                    "example": "ok"
                },
                "requestId": {
                    "description": "请求追踪号（X-Request-Id）",
                    "type": "string",
                    "format": "uuid",
                    "example": "01929d81-..."
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer \u003ctoken\u003e，登录后获得",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}